+++ src/NumericLiteral.cpp
@@ -1,10 +1,10 @@
 #include "NumericLiteral.h"
 
-Fraction::Fraction(int numerator, int denominator) {
+Fraction::Fraction(long numerator, long denominator) {
   SetFraction(numerator, denominator);
 }
 
-void Fraction::SetFraction(int numerator, int denominator) {
+void Fraction::SetFraction(long numerator, long denominator) {
   if (!denominator)
     throw(ComputerException("Dénominateur doit être non null."));
   numerator_ = numerator;
@@ -15,7 +15,7 @@ void Fraction::SetFraction(int numerator, int denominator) {
 void Fraction::Simplification() {
   if (!numerator_) denominator_ = 1;
 
-  int gcd = __gcd(numerator_, denominator_);
+  long gcd = __gcd(numerator_, denominator_);
   numerator_ /= gcd;
   denominator_ /= gcd;
 
@@ -24,11 +24,3 @@ void Fraction::Simplification() {
     numerator_ = -numerator_;
   }
 }
-
-// const QString Real::Print() const { 
-//   if(trunc(value_) == value_) {
-//     return QString::number(value_) + ".0";
-//   } else {
-//     return QString::number(value_); 
-//   }
-// }
